revise file layout and init logic for nvui
