+++ ui/analyse/src/view/nvuiView.ts
@@ -47,21 +47,21 @@ import { throttle } from 'lib/async';
 import { renderRetro } from '../retrospect/nvuiRetroView';
 import { playersView } from '../study/relay/relayPlayers';
 import { showInfo as tourOverview } from '../study/relay/relayTourView';
-import { NvuiContext } from '../analyse.nvui';
+import type { AnalyseNvuiContext } from '../analyse.nvui';
 
 const throttled = (sound: string) => throttle(100, () => site.sound.play(sound));
 const selectSound = throttled('select');
 const borderSound = throttled('outOfBound');
 const errorSound = throttled('error');
 
-export function initNvui({ ctrl, notify }: NvuiContext): void {
+export function initNvui({ ctrl, notify }: AnalyseNvuiContext): void {
   pubsub.on('analysis.server.progress', (data: AnalyseData) => {
     if (data.analysis && !data.analysis.partial) notify.set('Server-side analysis complete');
   });
   site.mousetrap.bind('c', () => notify.set(renderEvalAndDepth(ctrl))); // ? is 'c' for chat or eval?
 }
 
-export function renderNvui(ctx: NvuiContext): VNode {
+export function renderNvui(ctx: AnalyseNvuiContext): VNode {
   const { ctrl, deps, notify, moveStyle, pieceStyle, prefixStyle, positionStyle, boardStyle } = ctx;
   notify.redraw = ctrl.redraw;
   const d = ctrl.data,
@@ -232,7 +232,7 @@ export function clickHook(main?: (el: HTMLElement) => void, post?: () => void) {
   };
 }
 
-function boardEventsHook({ ctrl, pieceStyle, prefixStyle }: NvuiContext, el: HTMLElement): void {
+function boardEventsHook({ ctrl, pieceStyle, prefixStyle }: AnalyseNvuiContext, el: HTMLElement): void {
   const $board = $(el);
   const $buttons = $board.find('button');
   const steps = () => ctrl.tree.getNodeList(ctrl.path);
@@ -283,7 +283,7 @@ const noEvalStr = (ctrl: CevalCtrl) =>
         ? 'local evaluation not enabled'
         : '';
 
-function renderBestMove({ ctrl, moveStyle }: NvuiContext): string {
+function renderBestMove({ ctrl, moveStyle }: AnalyseNvuiContext): string {
   const noEvalMsg = noEvalStr(ctrl.ceval);
   if (noEvalMsg) return noEvalMsg;
   const node = ctrl.node,
@@ -312,7 +312,7 @@ function renderAriaResult(ctrl: AnalyseCtrl): VNode[] {
   ];
 }
 
-function renderCurrentLine({ ctrl, moveStyle }: NvuiContext) {
+function renderCurrentLine({ ctrl, moveStyle }: AnalyseNvuiContext) {
   if (ctrl.path.length === 0) return renderMainline(ctrl.mainline, ctrl.path, moveStyle.get(), !ctrl.retro);
   else {
     const futureNodes = ctrl.node.children.length > 0 ? ops.mainlineNodeList(ctrl.node.children[0]) : [];
@@ -320,7 +320,7 @@ function renderCurrentLine({ ctrl, moveStyle }: NvuiContext) {
   }
 }
 
-function onSubmit(ctx: NvuiContext, $input: Cash) {
+function onSubmit(ctx: AnalyseNvuiContext, $input: Cash) {
   const { ctrl, notify } = ctx;
   return (e: SubmitEvent) => {
     e.preventDefault();
@@ -346,7 +346,7 @@ type Command = 'p' | 's' | 'eval' | 'best' | 'prev' | 'next' | 'prev line' | 'ne
 type InputCommand = {
   cmd: Command;
   help: VNode | string;
-  cb: (ctrl: NvuiContext, input: string) => void;
+  cb: (ctrl: AnalyseNvuiContext, input: string) => void;
   invalid?: (ctrl: AnalyseCtrl) => boolean;
 };
 
@@ -433,7 +433,7 @@ function sendMove(uciOrDrop: string | DropMove, ctrl: AnalyseCtrl) {
   else if (ctrl.crazyValid(uciOrDrop.role, uciOrDrop.key)) ctrl.sendNewPiece(uciOrDrop.role, uciOrDrop.key);
 }
 
-function renderAcpl({ ctrl, moveStyle }: NvuiContext): LooseVNodes {
+function renderAcpl({ ctrl, moveStyle }: AnalyseNvuiContext): LooseVNodes {
   const analysis = ctrl.data.analysis;
   if (!analysis || ctrl.retro) return undefined;
   const analysisGlyphs = ['?!', '?', '??'];
@@ -470,7 +470,7 @@ function renderAcpl({ ctrl, moveStyle }: NvuiContext): LooseVNodes {
   return res;
 }
 
-const requestAnalysisBtn = ({ ctrl, notify, analysisInProgress }: NvuiContext) => {
+const requestAnalysisBtn = ({ ctrl, notify, analysisInProgress }: AnalyseNvuiContext) => {
   if (ctrl.ongoing || ctrl.synthetic || ctrl.hasFullComputerAnalysis()) return;
   return analysisInProgress()
     ? hl('p', 'Server-side analysis in progress')
@@ -503,7 +503,7 @@ function renderLineIndex(ctrl: AnalyseCtrl): string {
   return of > 1 ? `, line ${i + 1} of ${of} ,` : '';
 }
 
-function renderCurrentNode({ ctrl, moveStyle }: NvuiContext): string {
+function renderCurrentNode({ ctrl, moveStyle }: AnalyseNvuiContext): string {
   const node = ctrl.node;
   if (!node.san || !node.uci) return 'Initial position';
   return [
@@ -586,7 +586,7 @@ const redirectToSelectedHook = bind('change', (e: InputEvent) => {
   if (url) window.location.href = url;
 });
 
-function tourDetails({ ctrl, deps }: NvuiContext): VNode[] {
+function tourDetails({ ctrl, deps }: AnalyseNvuiContext): VNode[] {
   const ctx: RelayViewContext = { ...viewContext(ctrl, deps), allowVideo: false } as RelayViewContext;
   const tour = ctx.relay.data.tour;
   ctx.relay.redraw = ctrl.redraw;
