+++ ui/round/src/interfaces.ts
@@ -19,7 +19,7 @@ export interface NvuiPlugin {
   submitMove?: (submitStoredPremove?: boolean) => void;
   playPremove: (ctrl: RoundController) => void;
   premoveInput: string;
-  render(ctrl: RoundController): VNode;
+  render(): VNode;
 }
 
 export interface SocketOpts {
