+++ ui/round/src/round.ts
@@ -22,9 +22,7 @@ export async function initModule(opts: RoundOpts): Promise<RoundController> {
 }
 
 async function app(opts: RoundOpts): Promise<RoundController> {
-  const nvui = site.blindMode ? await site.asset.loadEsm<NvuiPlugin>('round.nvui') : undefined;
-  const ctrl = new RoundController(opts, redraw, nvui);
-
+  const ctrl = new RoundController(opts, redraw);
   const blueprint = view(ctrl);
   const el = (opts.element ?? document.querySelector('.round__app')) as HTMLElement;
